Move util methods to model
