+++ templates/single/quiz/parts/matching.php
@@ -1,7 +1,7 @@
 <div id="quiz-matching-ans-area" class="quiz-question-ans-choice-area tutor-mt-40 question-type-<?php echo $question_type; ?> <?php echo $answer_required ? 'quiz-answer-required' : ''; ?> ">
 	<div class="matching-quiz-question-desc tutor-d-flex tutor-align-start">
 		<?php
-			$rand_answers = tutor_utils()->get_answers_by_quiz_question( $question->question_id, true );
+			$rand_answers = \Tutor\Models\QuizModel::get_answers_by_quiz_question( $question->question_id, true );
 		foreach ( $rand_answers as $rand_answer ) { ?>
 			<div class="" style="display:flex; flex-direction: column; row-gap: 10px;">
 				<div class="tutor-draggable">
